extract doclet into main code
